day view fix and tidying routes
