+++ public/js/index.js
@@ -1,10 +1,11 @@
 $(function () {
 
     const time = $("#time");
+    time.text(dayjs().format("MM/DD/YYYY hh:mm A"));
 
     setInterval(function () {
-        time.text(dayjs().format("MM/DD/YYYY hh:mm:ss A"));
-    }, 1000);
+        time.text(dayjs().format("MM/DD/YYYY hh:mm A"));
+    }, 5000);
 
 
 });
