+++ views/layouts/main.handlebars
@@ -63,7 +63,7 @@
       {{{ body }}}
     </div>
   </section>
-  
+
 </body>
 
 </html>
