+++ views/task.handlebars
@@ -13,6 +13,7 @@
 </div>
 
 <script src="../js/deleteTask.js"></script>
+<script src="../js/index.js"></script>
